refactor: 잘못된 index.tsx 위치 수정
